separate non-32-bit-compatible code
