+++ appveyor.yml
@@ -13,8 +13,6 @@ environment:
   DATABASE: test
   GOVERSION: 113
   matrix:
-    - GOVERSION: 19
-      SQLINSTANCE: SQL2017
     - GOVERSION: 110
       SQLINSTANCE: SQL2017
     - GOVERSION: 111
@@ -40,6 +38,16 @@ environment:
     - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
       GOVERSION: 116
       SQLINSTANCE: SQL2017
+    - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+      GOVERSION: 117
+      SQLINSTANCE: SQL2017
+    - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+      GOVERSION: 118
+      SQLINSTANCE: SQL2017
+    - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+      GOVERSION: 118-x86
+      SQLINSTANCE: SQL2017
+      GOARCH: 386
 
 install:
   - set GOROOT=c:\go%GOVERSION%
