+++ bulkcopy_test.go
@@ -1,5 +1,5 @@
-//go:build go1.9
-// +build go1.9
+//go:build go1.9 && !386
+// +build go1.9,!386
 
 package mssql
